+++ apps/functions-subscription/src/adapters/azure/functions/__tests__/mocks.ts
@@ -11,9 +11,11 @@ export const makeFunctionContext = () =>
 export const makeTestSystemEnv = () => {
   const insertSubscriptionMock = vi.fn();
   const processSubscriptionRequestMock = vi.fn();
+  const getSubscriptionMock = vi.fn();
   return {
     insertSubscription: insertSubscriptionMock,
     processSubscriptionRequest: processSubscriptionRequestMock,
+    getSubscription: getSubscriptionMock,
   };
 };