+++ apps/functions-subscription/src/config.ts
@@ -5,22 +5,22 @@ import * as E from 'fp-ts/lib/Either';
 import { NonEmptyString } from '@pagopa/ts-commons/lib/strings';
 
 export interface Config {
-  readonly subscriptionRequest: {
-    readonly eventhub: {
-      readonly connectionString: string;
-      readonly name: string;
+  readonly eventhubs: {
+    readonly namespace: string;
+    readonly names: {
+      readonly subscriptionRequest: string;
     };
   };
   readonly cosmosdb: {
-    readonly connectionString: string;
+    readonly endpoint: string;
     readonly databaseName: string;
   };
 }
 
 const EnvsCodec = t.strict({
-  COSMOSDB_CONNECTION_STRING: NonEmptyString,
+  COSMOSDB_ENDPOINT: NonEmptyString,
   COSMOSDB_DATABASE_NAME: NonEmptyString,
-  SUBSCRIPTION_REQUEST_EVENTHUB_CONNECTION_STRING: NonEmptyString,
+  EVENTHUB_NAMESPACE: NonEmptyString,
   SUBSCRIPTION_REQUEST_EVENTHUB_NAME: NonEmptyString,
 });
 
@@ -32,16 +32,15 @@ export const parseConfig = (
     E.bimap(
       (errors) => PR.failure(errors).join('\n'),
       (envs) => ({
-        cosmosdb: {
-          connectionString: envs.COSMOSDB_CONNECTION_STRING,
-          databaseName: envs.COSMOSDB_DATABASE_NAME,
-        },
-        subscriptionRequest: {
-          eventhub: {
-            connectionString:
-              envs.SUBSCRIPTION_REQUEST_EVENTHUB_CONNECTION_STRING,
-            name: envs.SUBSCRIPTION_REQUEST_EVENTHUB_NAME,
+        eventhubs: {
+          namespace: envs.EVENTHUB_NAMESPACE,
+          names: {
+            subscriptionRequest: envs.SUBSCRIPTION_REQUEST_EVENTHUB_NAME,
           },
+        },
+        cosmosdb: {
+          endpoint: envs.COSMOSDB_ENDPOINT,
+          databaseName: envs.COSMOSDB_DATABASE_NAME,
         },
       }),
     ),