+++ infra/resources/modules/commons/locals.tf
@@ -4,4 +4,3 @@ locals {
   is_prod              = var.env_short == "p" ? true : false
   application_basename = "trial-system"
 }
-