+++ infra/resources/modules/commons/network.tf
@@ -89,26 +89,6 @@ module "pendpoints_snet" {
 
 }
 
-resource "azurerm_private_endpoint" "sql" {
-
-  name                = format("%s-private-endpoint-sql-01", local.project)
-  location            = var.location
-  resource_group_name = azurerm_resource_group.net_rg.name
-  subnet_id           = module.pendpoints_snet.id
-
-  private_service_connection {
-    name                           = format("%s-private-endpoint-sql", local.project)
-    private_connection_resource_id = module.cosmosdb_account.id
-    is_manual_connection           = false
-    subresource_names              = ["Sql"]
-  }
-
-  private_dns_zone_group {
-    name                 = "private-dns-zone-group"
-    private_dns_zone_ids = [azurerm_private_dns_zone.privatelink_documents.id]
-  }
-}
-
 resource "azurerm_private_dns_zone" "privatelink_azure_websites" {
   name                = "privatelink.azurewebsites.net"
   resource_group_name = azurerm_resource_group.net_rg.name
@@ -118,6 +98,19 @@ resource "azurerm_private_dns_zone_virtual_network_link" "websites_link" {
   name                  = azurerm_virtual_network.vnet.name
   resource_group_name   = azurerm_resource_group.net_rg.name
   private_dns_zone_name = azurerm_private_dns_zone.privatelink_azure_websites.name
+  virtual_network_id    = azurerm_virtual_network.vnet.id
+}
+
+resource "azurerm_private_dns_zone" "privatelink_servicebus" {
+  name                = "privatelink.servicebus.windows.net"
+  resource_group_name = azurerm_resource_group.net_rg.name
+
+}
+
+resource "azurerm_private_dns_zone_virtual_network_link" "evh_link" {
+  name                  = azurerm_virtual_network.vnet.name
+  resource_group_name   = azurerm_resource_group.net_rg.name
+  private_dns_zone_name = azurerm_private_dns_zone.privatelink_servicebus.name
   virtual_network_id    = azurerm_virtual_network.vnet.id
 }