+++ apps/functions-subscription/CHANGELOG.md
@@ -1,5 +1,16 @@
 # functions-subscription
 
+## 0.2.0
+
+### Minor Changes
+
+- 2739a3f: Add get subscription endpoint
+
+### Patch Changes
+
+- d7a1f73: Change authLevel for create subscription endpoint
+- 7107ec4: [IOPLT-534] Migrate to managed identities
+
 ## 0.1.0
 
 ### Minor Changes